requirements.txt
openai
faiss-cpu
numpy
python-dotenv
gradio
prefect